:recycle: Remove authOptions from api/auth and create a server auth helper
