+++ app/dashboard/layout.tsx
@@ -1,12 +1,10 @@
-import { authOptions } from '@/api/auth/[...nextauth]/route';
 import SessionExpiredDialog from '@/auth/components/session-expired-dialog';
+import { auth } from '@/auth/lib/auth';
 import Header from '@/components/layout/header';
 import Sidebar from '@/components/layout/sidebar';
-import { getServerSession } from 'next-auth';
-//import { signOut } from 'next-auth/react';
 
 export default async function DashboardLayout({ children }: { children: React.ReactNode }) {
-  const session = await getServerSession(authOptions);
+  const session = await auth();
 
   const sessionExpired = session?.error === 'RefreshTokenExpired';
